+++ src/main/java/org/example/Producer/CfgProduceGraphic.java
@@ -11,7 +11,7 @@ import java.util.List;
 @XmlRootElement(name = "graphic")
 public class CfgProduceGraphic {
     @Getter
-    private List<Double> time;
+    private List<Integer> time;
     @Getter
     private List<Double> power;
     @Getter
